PrinterSchemes/requirements.txt
aiohttp==3.7.4
astor==0.8.1
autopep8==1.5.7
axelrod==4.13.0
big-sleep==0.8.5
blinker==1.4
brotlipy==0.7.0
charset-normalizer==2.0.3
coverage==5.5
cups==0.0.6
cython==0.29.24
deep-daze==0.10.2
django-allauth==0.45.0
django-countries==7.2.1
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
facebook-sdk==3.1.0
fbchat==1.9.7
ffmpeg==1.4
flask-ngrok==0.0.25
flask-ngrok2==0.2.4
gametheory==0.0.0
gym==0.18.3
jaraco.text==3.12.1
jupyterlab==3.2.9
keras-applications==1.0.8
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
nashpy==0.0.41
ngrok==1.4.0
olefile==0.46
pep8==1.7.1
pip-chill==1.0.3
platformdirs==4.2.2
pyaudio==0.2.11
pyngrok==7.2.0
pyopenssl==20.0.1
pyqt5==5.15.4
pyreadline==2.1
pysocks==1.7.1
python-decouple==3.4
python-dotenv==0.21.0
quantecon==0.7.2
scikit-learn==0.24.2
sentencepiece==0.1.96
squareup==38.0.0.20240821
stripe==2.60.0
subprocess.run==0.0.8
tensorflow==2.5.0
tensorflow-gpu==2.5.0
tomli==2.0.1
torchaudio==0.9.0
transformers==4.9.1
twilio==7.14.0
vosk==0.3.30
win-inet-pton==1.1.0
wincertstore==0.2
woocommerce==3.0.0
